arm64: dts: qcom: Honor Magicbook: Add FSUSB42 USB switches
The setup is the same on the Honor Magicbook as on the QCP, so use the
same setup, based on this patch:

Unlike most X1E boards, the QCP does not have Parade PS8830 retimers on
the three USB-C ports. Instead, there are FSUSB42 USB switches for each
port that handle orientation switching for the SBU lines. The overall
setup is similar to the gpio-sbu-mux defined for sc8280xp-crd and the
ThinkPad X13s.